refactor: move generic api tools out of the kitsu module
Some items such as the AuthProvider and the Auth interceptors are not specific to Kitsu but they
need to be used within Auth. Move them out of the Kitsu package and move into the Auth package so
they could be reused with any other apis
