+++ libraries/datasource/auth/src/main/java/com/chesire/nekome/datasource/auth/remote/AuthInjectionInterceptor.kt
@@ -1,16 +1,17 @@
-package com.chesire.nekome.kitsu.interceptors
+package com.chesire.nekome.datasource.auth.remote
 
-import com.chesire.nekome.kitsu.AuthProvider
+import com.chesire.nekome.datasource.auth.local.AuthProvider
 import okhttp3.Interceptor
 import okhttp3.Response
 import javax.inject.Inject
 
 /**
- * Interceptor to push the authorization header into Kitsu requests.
+ * Interceptor to push the authorization header into api requests.
  */
 class AuthInjectionInterceptor @Inject constructor(
     private val provider: AuthProvider
 ) : Interceptor {
+
     override fun intercept(chain: Interceptor.Chain): Response {
         val authenticatedRequest = chain.request()
             .newBuilder()
