+++ libraries/kitsu/src/main/java/com/chesire/nekome/kitsu/ResponseParsing.kt
@@ -1,6 +1,7 @@
 package com.chesire.nekome.kitsu
 
 import com.chesire.nekome.core.Resource
+import com.chesire.nekome.datasource.auth.AuthException
 import retrofit2.Response
 import java.net.UnknownHostException
